learning_system/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
pytz==2024.1
tzdata==2024.1

